+++ app/(team)/[id]/edit/_components/EditForm.tsx
@@ -7,20 +7,21 @@ import Button from "@/app/_components/Button";
 import Form from "@/app/_components/Form";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import Image from "next/image";
-import { useParams } from "next/navigation";
+import { useParams, useRouter } from "next/navigation";
 import { useCallback } from "react";
 
 type FormContext = Parameters<Parameters<typeof Form>[0]["onSubmit"]>[0];
 
 export default function TeamEditForm() {
 	const queryClient = useQueryClient();
+	const router = useRouter();
 
 	const { id } = useParams<{ id: string }>();
 
-	const teamInfo = useQuery({
+	const { data: teamInfo } = useQuery({
 		queryKey: ["teamInfo", { id: Number(id) }],
 		queryFn: async () => {
-			const response = await API["{teamId}/groups/{id}"].GET({ id: Number(id) }).then((res) => res);
+			const response = await API["{teamId}/groups/{id}"].GET({ id: Number(id) });
 			return response;
 		},
 	});
@@ -49,16 +50,14 @@ export default function TeamEditForm() {
 			const response = await API["{teamId}/groups/{id}"].PATCH({ id: Number(id) }, payload);
 			return response;
 		},
-		onSuccess: (data) => {
-			// TODO: 팀 생성 성공 시 팀 페이지로 이동
-			console.log("Success: ", data);
-
+		onSuccess: () => {
 			// 쿼리 무효화
 			queryClient.invalidateQueries({ queryKey: ["teamInfo"] });
+
+			router.push(`/${id}`);
 		},
-		onError: (error) => {
-			console.log("Error: ", error);
-			// TODO: 팀 생성 실패 시 에러 메시지 표시
+		onError: (error, ctx) => {
+			ctx.setError("teamName", "팀 수정에 실패했습니다.");
 		},
 	});
 
@@ -72,76 +71,74 @@ export default function TeamEditForm() {
 	);
 
 	return (
-		<main className="h-dvh min-w-[320px] bg-background-primary text-text-default">
-			<section className="flex size-full flex-col items-center justify-center">
-				<div className="pb-20">
-					<h1 className="text-[40px] font-medium leading-[48px] text-white">팀 수정하기</h1>
-				</div>
+		<section className="flex size-full flex-col items-center justify-center">
+			<div className="pb-20">
+				<h1 className="text-[40px] font-medium leading-[48px] text-white">팀 수정하기</h1>
+			</div>
 
-				<div className="w-full">
-					<Form onSubmit={handleTeamManagement}>
-						<div className="mx-auto flex size-full max-w-[340px] flex-col tablet:max-w-[460px]">
-							<div className="w-fit">
-								<label htmlFor="team-name" className="w-full text-start text-text-primary">
-									팀 프로필
-								</label>
-								<div className="pb-3" />
-								<Form.ImageInput id="profileImage" tests={[{ type: "file_size", data: 1048576, error: "이미지 파일 크기는 1MB 이하여야 합니다" }]}>
-									{(file) => (
-										// eslint-disable-next-line react/jsx-no-useless-fragment
-										<>
-											{file || teamInfo.data?.image ? (
-												<div className="relative flex size-16 items-center justify-center rounded-[12px] border-2 border-border-primary/10">
-													<Image
-														src={(file as string) ?? teamInfo.data?.image ?? ""}
-														alt="Profile Preview"
-														fill
-														className="rounded-[12px] object-cover object-center"
-													/>
-													<div className="relative size-full">
-														<Image src="/icons/editIcon.svg" alt="Profile Preview" width={20} height={20} className="absolute -bottom-2 -right-2" />
-													</div>
-												</div>
-											) : (
-												<div className="relative flex size-16 items-center justify-center rounded-[12px] border-2 border-border-primary/10 bg-background-secondary">
-													<div className="relative size-5">
-														<Image src="/icons/imageIcon.svg" alt="Profile Preview" fill />
-													</div>
-
+			<div className="w-full">
+				<Form onSubmit={handleTeamManagement}>
+					<div className="mx-auto flex size-full max-w-[340px] flex-col tablet:max-w-[460px]">
+						<div className="w-fit">
+							<label htmlFor="team-name" className="w-full text-start text-text-primary">
+								팀 프로필
+							</label>
+							<div className="pb-3" />
+							<Form.ImageInput id="profileImage" tests={[{ type: "file_size", data: 1048576, error: "이미지 파일 크기는 1MB 이하여야 합니다" }]}>
+								{(file) => (
+									// eslint-disable-next-line react/jsx-no-useless-fragment
+									<>
+										{file || teamInfo?.image ? (
+											<div className="relative flex size-16 items-center justify-center rounded-[12px] border-2 border-border-primary/10">
+												<Image
+													src={(file as string) ?? teamInfo?.image ?? ""}
+													alt="Profile Preview"
+													fill
+													className="rounded-[12px] object-cover object-center"
+												/>
+												<div className="relative size-full">
 													<Image src="/icons/editIcon.svg" alt="Profile Preview" width={20} height={20} className="absolute -bottom-2 -right-2" />
 												</div>
-											)}
-										</>
-									)}
-								</Form.ImageInput>
-							</div>
-							<Form.Error htmlFor="team-image" />
+											</div>
+										) : (
+											<div className="relative flex size-16 items-center justify-center rounded-[12px] border-2 border-border-primary/10 bg-background-secondary">
+												<div className="relative size-5">
+													<Image src="/icons/imageIcon.svg" alt="Profile Preview" fill />
+												</div>
 
-							<div className="pt-6" />
-
-							<label htmlFor="team-name" className="w-full pb-3 text-start text-text-primary">
-								팀 이름
-							</label>
-							<Form.Input
-								id="teamName"
-								type="text"
-								placeholder="팀 이름을 입력하세요"
-								init={teamInfo.data?.name}
-								tests={[{ type: "require", data: true, error: "팀 이름은 필수입니다" }]}
-							/>
-							<Form.Error htmlFor="team-name" />
-
-							<div className="pt-10" />
-
-							<div className="h-12">{teamEditMutation.status === "pending" ? <Button disabled>팀 수정 중</Button> : <Form.Submit>수정하기</Form.Submit>}</div>
+												<Image src="/icons/editIcon.svg" alt="Profile Preview" width={20} height={20} className="absolute -bottom-2 -right-2" />
+											</div>
+										)}
+									</>
+								)}
+							</Form.ImageInput>
 						</div>
-					</Form>
-				</div>
+						<Form.Error htmlFor="team-image" />
 
-				<div className="max-w-[340px] pt-6 tablet:max-w-[460px]">
-					<p className="text-md font-normal text-text-primary tablet:text-lg">팀 이름은 회사명이나 모임 이름 등으로 설정하면 좋아요.</p>
-				</div>
-			</section>
-		</main>
+						<div className="pt-6" />
+
+						<label htmlFor="team-name" className="w-full pb-3 text-start text-text-primary">
+							팀 이름
+						</label>
+						<Form.Input
+							id="teamName"
+							type="text"
+							placeholder="팀 이름을 입력하세요"
+							init={teamInfo?.name}
+							tests={[{ type: "require", data: true, error: "팀 이름은 필수입니다" }]}
+						/>
+						<Form.Error htmlFor="team-name" />
+
+						<div className="pt-10" />
+
+						<div className="h-12">{teamEditMutation.status === "pending" ? <Button disabled>팀 수정 중</Button> : <Form.Submit>수정하기</Form.Submit>}</div>
+					</div>
+				</Form>
+			</div>
+
+			<div className="max-w-[340px] pt-6 tablet:max-w-[460px]">
+				<p className="text-md font-normal text-text-primary tablet:text-lg">팀 이름은 회사명이나 모임 이름 등으로 설정하면 좋아요.</p>
+			</div>
+		</section>
 	);
 }
