+++ app/(team)/join-team/layout.tsx
