+++ app/(team)/layout.tsx
@@ -1,0 +1,3 @@
+export default function JoinTeamLayout({ children }: Readonly<React.PropsWithChildren>) {
+	return <main className="h-dvh min-w-[320px] bg-background-primary text-text-default">{children}</main>;
+}
