+++ app/_components/Form.tsx
@@ -241,7 +241,7 @@ Form.Input = function Input({
 };
 
 Form.TextArea = function TextArea({ id, init, tests, placeholder }: Readonly<{ id: string; init?: string; tests?: Validator[]; placeholder?: string }>) {
-	const ctx = uinseCTX();
+	const ctx = useCTX();
 
 	const [value, setValue] = useState("");
 	const [focus, setFocus] = useState(false);
